ft_atoi: test ok, execution: ok   *corrigido bug na funcao strnstr

+++ ft_bzero.c
@@ -1,6 +1,6 @@
-
-void *ft_bzero(void *BLOCK, unsigned int SIZE) {
-  unsigned int index; 
+#include "./libft.h"
+void *ft_bzero(void *BLOCK, size_t SIZE) {
+  size_t index; 
 
   index = 0; 
   while (index < SIZE) {
